requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
requests
geopy
lxml
aioredis
haversine
openrouteservice
httpx
redis
fastapi_limiter
pytest